Designate KP.4 (JN.1.11.1.4, C6070T, India), KP.4.1 (S:R346T, C19884T), KP.4.2 (S:R346T, S:K187R)
Note: I'm only designating this now because there was uncertainty whether 6070T happened
before or after the 346T. The 4.1 and 4.2 could also be sublineages of KP.2
After all, the fact that there's little 6070T with just 346T and that most have
an extra mutation (C19884T for KP.4.1, and S:K187R for KP.4.2) made me content
with making this a convergent lineage.

We likely won't ever know what the "truth" is but that's ok, Pango is not
perfect and doesn't have to be for it to be useful.